# backend/openapi/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    ChatMessage:
      type: object
      properties:
        role:
          type: string
        content:
          type: string
        cached:
          type: boolean
...
    ThreeScaleSourceStatus:
      type: object
      description: Reachability status for a configured 3scale Admin API source
      properties:
        id:
          type: string
        label:
          type: string
        adminUrl:
          type: string
        configured:
          type: boolean
        enabled:
          type: boolean
        reachable:
          type: boolean
        error:
          type: string
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      description: Authentication
paths:
  /api/apicast/analyze/{namespace}/{name}:
    get:
      operationId: analyzeApicast
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Analyze
      tags:
      - API Cast Resource
...
  /api/apicast/discover/{namespace}:
    get:
      operationId: discoverByNamespace
      parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Discover By Namespace
      tags:
      - API Cast Resource
  /api/apicast/map:
    post:
      operationId: mapAPICastToIstio
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Map API Cast To Istio
      tags:
      - API Cast Resource
      security:
      - SecurityScheme:
        - admin
  /api/apicast/map-all:
    post:
      operationId: mapAll
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Map All
      tags:
      - API Cast Resource
      security:
      - SecurityScheme:
        - admin
...
  /api/chat:
    post:
      operationId: chat
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessage"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Chat
      tags:
      - Chat Resource
      security:
      - SecurityScheme:
        - admin
...
  /api/chat/warm-faq:
    post:
      operationId: warmFaqCache
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Warm Faq Cache
      tags:
      - Chat Resource
      security:
      - SecurityScheme:
        - admin
...
  /api/threescale/sources/{id}/status:
    get:
      operationId: getSourceStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreeScaleSourceStatus"
      summary: Get Source Status
      tags:
      - Three Scale Resource
...
info:
  title: ApiShift API
  version: 0.3.0
  description: 3scale to Connectivity Link migration console REST API
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value